Move menu preview to new component
